requirements.txt
werkzeug==2.3.7
Flask
flask_sqlalchemy
flask_login
flask_wtf
WTForms
WTForms[email]
flask-bcrypt
pyotp
flask-qrcode
python-dotenv
flask-talisman
pyopenssl